Add test of an action requiring XML escaped characters
Have some tests we expect to fail, flag them as such.

All the places that were running a test that expected errors had some
lines of code to warn the user that errors were expected. Since the
run_etch method knows if errors are expected I've centralized the
printing of that warning there, DRYing things up a bit.