Added authentication support. The client will sign messages using its ssh host key.
Add support for local requests.

Update the method used to eliminate the OpenSSL "using default DH
parameters" warning.
http://techbits.blogspot.com/2009/08/ive-been-using-workaround-shown-at.html

Add timeout to output capturing so that ill-behaved daemons don't cause
etch to hang around forever.  Add unit tests for output capturing.

Note that the output capturing timeout test is commented out as the
timeout is rather long, thus causing the test to take a long time to
run.  Not quite sure the best way to handle that.

Move Etch::Client.initialize arguments to a hash so that the method
doesn't take 10 or whatever arguments.

Set etch version as a constant in the library (Etch::Client::VERSION).

Add a command line option to etch to display the version.

Modify the make targets to insert the version into the appropriate place
rather than having the version hard-coded in a bunch of files.

Add redhatprep, debianprep and solarisprep targets which install the
packages necessary to extract the version from the library file.

Add code to require rubygems if requiring facter fails.

Change /usr/lib/ruby/site_ruby to /usr/local/lib/ruby/site_ruby for the
Debian package.  That seems to be the correct location on Debian
systems.