Add support for guarded commands http://sourceforge.net/apps/trac/etch/wiki/ConfigurationCommands
Add local mode so that the etch client can work against a local config
repository rather than a server.

Rename etch.rb to etchclient.rb to reflect its purpose.

Break out the code from etchserver.rb which is not server-specific to a
separate file named etch.rb. Abstract the XML calls in that file so that
either REXML or LibXML can be used. This will support future
modifications to the client to allow the client to run in standalone
mode against a local configuration repository.

Abstract the XML response generation in Etch::Server so that the server
does not depend on LibXML. All code should now fall back to REXML if
LibXML is not available.

Add in support for an xmllib environment variable that can be used to
force the code to a particular XML library.

Fix bug in configfilter! which was triggering a bug in LibXML

Fix issues with timing around saving originals, for non-directories etch
was saving the original the first time it was asked for the original
contents or metadata from the server.

Fix bug with ordering of creating backups and creating directory
structure.