Add support for requesting specific commands.
Add featuring allowing the server to tell the client to retry certain
commands on the next request. The "need sum" mechanism is used to do the
same thing for files, but as commands have no original content saving
mechanism we have to have an explicit feature for retries. The server
needs this if it is unable to generate all of the prerequisites for the
command, usually because the command depends on a file for which the
server needs original content info.