Explicitly invoke user actions via "sh -c" so that syntax like "FOO=bar myprogram" works
When generating commands remove any steps where the guard and command 
have been filtered out.  Raise an exception if either the guard or 
command were filtered out but not both.  Improve error message in all 
cases to try to help the user track down the offending entry.

Modify check_for_inconsistency to allow for 0 elements.

When preping a blank local request in the client set up the :commands
hash if it will be needed.