In the test_nested_target method in test_actions.rb there was a section of code which called Tempfile.new('etchtest').path, then immediately deleted the file and replaced it with a directory.  The Tempfile object would get garbage collected later and Tempfile would throw an error attempting to delete what was now a directory.  Replace that with a call to deleted_tempfile so that Tempfile doesn't try to clean things up later.
Replace a bunch of instances of Tempfile.new('etchtest').path with a new
method released_tempfile, which explicitly does a close! on the Tempfile
and then recreates the file.  It may be overly paranoid of me, but it
seems that depending on Ruby to more or less immediately garbage collect
the Tempfile object is not ideal.