Replace open flash chart graphs on dashboard with flot graph.  Yay, no more flash.
I never liked the way the dashboard looked with the status pie chart and
the client count chart, and it looked even worse with the flot graphs.
The status pie chart was a bit superfluous given that the same data is
right there in the Client Summary.   So I've commented it out for now.
If someone misses it we'll figure out a way to display it without
burning anyone's eyes.  Otherwise we can toast the commented out bits at
a later time.