chdir back to / after processing individual files or commands.
We chdir into the directory for the individual file/command before
processing it so that the user can refer to other files in the same
directory using relative paths.  In practice it hasn't caused any
problems to leave the server chdir'd into that directory until it
processes the next file, but it does generate sporadic errors from the
test suite when the directory disappears as the repository is updated.
So to be on the safe and robust side we now chdir back to / after
processing an individual file or command.