Convert config column to blob, compress contents
Text columns in MySQL are limited to 64k.  I have config files that are
larger.  It apparently was silently truncating data prior to rails 4.
With rails 4 it was explicitly failing.  The data in the config column
is currently never searched, it may well be binary data anyway, etc.  So
converting it to a blob makes sense.  And since that data is often
compressible text go ahead and compress it to save some space.