Support YAML format config, commands  Issue #20
Add support for over-the-wire client/server YAML format.  XML is also
still supported.  Configs read from either YAML or XML are converted to
a unified in-memory format that can be serialized to XML or YAML for
clients, retaining backwards compatibility with old (XML-only) clients.

Local requests no longer need be XML.  They are provided to external
sources via @local_requests as before as an XML document, but are now
also provided via @local_requests_array as an array of strings as read
from the files on the client.  You'll probably want to ensure that all
scripts are converted to use @local_requests_array before creating any
local requests files in non-XML formats.  Otherwise the concatenated
files in @local_requests will be hard to distinguish.