Make hosts.yml optional.  Load DTD files on demand.
Trying to reduce the number of magic files required to get an etch
server running.  For some historical reason I don't even remember at
this point hosts.xml was required to exist even if it was basically
empty.  I no longer see a reason to require that.  And the DTD files are
no longer needed if you have a pure YAML configuration, so load them
only if we encounter a XML config file.